refactor data provider to grouped-row-data-provider for grouped-row-sort tests
